# openapi.yaml
openapi: 3.1.0
info:
  title: Waternity API
  description: |
    Waternity is a transparent water sales and settlement platform built on Hedera Hashgraph.
    This API provides endpoints for managing water projects, wells, settlements, and audit trails.
    
    ## Authentication
    The API uses cookie-based authentication with httpOnly JWT tokens. After successful login,
    a secure session cookie (`wty_sess`) is set with 8-hour expiry.
    
    ## Role-Based Access Control
    - **INVESTOR**: Can invest in projects and view settlements
    - **OPERATOR**: Can create wells/projects, manage operations, control meters/valves
    - **AGENT**: Can verify/approve settlements and anchor documents
    - **ADMIN**: Full system access
    
    ## Hedera Integration
    All operations are recorded on Hedera Consensus Service (HCS) for immutable audit trails.
    Links to HashScan explorer and Mirror Node are provided for verification.
  version: 1.0.0
  contact:
    name: Waternity Team
    url: https://waternity.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://waternity.vercel.app
    description: Production server
paths:
...
  /api/hcs/events:
    post:
      tags:
        - HCS Events
      summary: Submit HCS event
      description: |
        Submit an event to Hedera Consensus Service for immutable audit trail.
        Events are validated against JSON Schema 2020-12 specifications.
      operationId: submitHcsEvent
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HcsEventRequest'
            examples:
              projectCreated:
                summary: Project created event
                value:
                  messageId: "550e8400-e29b-41d4-a716-446655440000"
                  eventType: "project_created"
                  timestamp: "2024-01-15T10:30:00Z"
                  triggeredBy: "0.0.12345"
                  entityId: "proj_001"
                  entityType: "project"
                  eventData:
                    name: "Sumur Air Bersih Desa Makmur"
                    description: "Proyek pembangunan sumur air bersih untuk 500 keluarga"
                    fundingTarget: 10000
                    location:
                      address: "Desa Makmur, Kecamatan Sejahtera"
                      latitude: -6.2088
                      longitude: 106.8456
      responses:
        '201':
          description: Event submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HcsEventResponse'
              examples:
                success:
                  value:
                    success: true
                    messageId: "550e8400-e29b-41d4-a716-446655440000"
                    transactionId: "0.0.12345@1705315800.123456789"
                    consensusTimestamp: "1705315800.123456789"
                    links:
                      hashscan: "https://hashscan.io/testnet/transaction/0.0.12345@1705315800.123456789"
                      mirror: "https://testnet.mirrornode.hedera.com/api/v1/transactions/0.0.12345@1705315800.123456789"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'

  # Wells
...
  /api/settlements/request:
    post:
      tags:
        - Settlements
      summary: Request settlement
      description: |
        Request a new settlement for revenue distribution.
        Only operators can request settlements.
      operationId: requestSettlement
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettlementRequest'
      responses:
        '201':
          description: Settlement request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/settlements/execute:
    post:
      tags:
        - Settlements
      summary: Execute settlement
      description: |
        Execute an approved settlement by transferring tokens.
        Only agents can execute settlements.
      operationId: executeSettlement
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettlementExecutionRequest'
      responses:
        '200':
          description: Settlement executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'

  # Documents
...
  /api/meta/links:
    get:
      tags:
        - Metadata
      summary: Get verification links
      description: |
        Get HashScan and Mirror Node links for verification.
        This endpoint is publicly accessible for transparency.
      operationId: getVerificationLinks
      parameters:
        - name: wellId
          in: query
          required: true
          description: Well ID to get links for
          schema:
            type: string
            example: "well_001"
      responses:
        '200':
          description: Verification links retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationLinksResponse'
              examples:
                success:
                  value:
                    hashscan:
                      topic: "https://hashscan.io/testnet/topic/0.0.789012"
                      token: "https://hashscan.io/testnet/token/0.0.123456"
                      file: "https://hashscan.io/testnet/file/0.0.345678"
                    mirror:
                      topic: "https://testnet.mirrornode.hedera.com/api/v1/topics/0.0.789012/messages"
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'

  # Health
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: wty_sess
      description: |
        Session-based authentication using httpOnly JWT cookies.
        The cookie is automatically set after successful login and cleared on logout.
        
        Cookie attributes:
        - HttpOnly: true (prevents XSS attacks)
        - Secure: true (HTTPS only in production)
        - SameSite: Lax (CSRF protection)
        - Max-Age: 28800 (8 hours)
  schemas:
...
    HcsEventRequest:
      $ref: 'hcs-event-schemas.json#/definitions/baseEvent'
    HcsEventResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        messageId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        transactionId:
          type: string
          description: Hedera transaction ID
          example: "0.0.12345@1705315800.123456789"
        consensusTimestamp:
          type: string
          description: Consensus timestamp from Hedera
          example: "1705315800.123456789"
        links:
          $ref: '#/components/schemas/HederaLinks'

    # Well Schemas
...
    SettlementRequest:
      type: object
      required:
        - messageId
        - wellId
        - periodStart
        - periodEnd
        - totalRevenue
      properties:
        messageId:
          type: string
          format: uuid
          description: Idempotency key
        wellId:
          type: string
          example: "well_001"
        periodStart:
          type: string
          format: date
          example: "2024-01-01"
        periodEnd:
          type: string
          format: date
          example: "2024-01-31"
        totalRevenue:
          type: number
          minimum: 0
          example: 5000.50
        description:
          type: string
          example: "Monthly settlement for January 2024"
...
    SettlementExecutionRequest:
      type: object
      required:
        - messageId
        - settlementId
      properties:
        messageId:
          type: string
          format: uuid
        settlementId:
          type: string
          example: "settlement_001"
    SettlementResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        settlement:
          $ref: '#/components/schemas/Settlement'
        links:
          $ref: '#/components/schemas/HederaLinks'
    Settlement:
      type: object
      properties:
        id:
          type: string
          example: "settlement_001"
        wellId:
          type: string
          example: "well_001"
        status:
          type: string
          enum: [pending, approved, executed, rejected]
          example: "approved"
        totalRevenue:
          type: number
          example: 5000.50
        payouts:
          type: array
          items:
            $ref: '#/components/schemas/Payout'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Payout:
      type: object
      properties:
        accountId:
          type: string
          pattern: "^0\\.0\\.[0-9]+$"
          example: "0.0.12345"
        amount:
          type: number
          minimum: 0
          example: 1250.25
        role:
          type: string
          example: "investor"
        percentage:
          type: number
          minimum: 0
          maximum: 100
          example: 25.0

    # Document Schemas
...
    HederaLinks:
      type: object
      properties:
        hashscan:
          type: string
          format: uri
          description: HashScan explorer link
          example: "https://hashscan.io/testnet/topic/0.0.789012"
        mirror:
          type: string
          format: uri
          description: Mirror Node API link
          example: "https://testnet.mirrornode.hedera.com/api/v1/topics/0.0.789012/messages"
    VerificationLinksResponse:
      type: object
      properties:
        hashscan:
          type: object
          properties:
            topic:
              type: string
              format: uri
              example: "https://hashscan.io/testnet/topic/0.0.789012"
            token:
              type: string
              format: uri
              example: "https://hashscan.io/testnet/token/0.0.123456"
            file:
              type: string
              format: uri
              example: "https://hashscan.io/testnet/file/0.0.345678"
        mirror:
          type: object
          properties:
            topic:
              type: string
              format: uri
              example: "https://testnet.mirrornode.hedera.com/api/v1/topics/0.0.789012/messages"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              value:
                error: "validation_error"
                message: "Request validation failed"
                details:
                  field: "username"
                  issue: "Username is required"
                timestamp: "2024-01-15T10:30:00Z"
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                error: "unauthorized"
                message: "Authentication required"
                timestamp: "2024-01-15T10:30:00Z"
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  needRole:
                    type: string
                    description: Required role for this operation
          examples:
            forbidden:
              value:
                error: "forbidden"
                message: "Insufficient permissions"
                needRole: "OPERATOR"
                timestamp: "2024-01-15T10:30:00Z"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                error: "not_found"
                message: "Resource not found"
                timestamp: "2024-01-15T10:30:00Z"
    ConflictError:
      description: Resource conflict (e.g., duplicate operation)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            conflict:
              value:
                error: "conflict"
                message: "Operation already exists"
                details:
                  messageId: "550e8400-e29b-41d4-a716-446655440000"
                timestamp: "2024-01-15T10:30:00Z"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal:
              value:
                error: "internal_error"
                message: "An internal error occurred"
                timestamp: "2024-01-15T10:30:00Z"
...